+++ docs/changelog.md
@@ -2,3 +2,12 @@
 title: 更新日志
 sidebar_position: 2
 ---
+
+## 2023-05-21
+
+- 增加 Awesomes，整理了一些目前看到的 Awesome 仓库
+- 增加 我的LLM项目
+
+## 2023-05-20
+
+项目上线。